+++ src/main/java/com/cloudogu/smeagol/ServletInitializer.java
@@ -6,10 +6,8 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
  * Allows for being deployed as a war into an external container
  */
 public class ServletInitializer extends SpringBootServletInitializer {
-
     @Override
     protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
         return application.sources(Smeagol.class);
     }
-
 }